Macroprosody analysis

+++ pncore/datastore/AnnotationDatastore.h
@@ -173,7 +173,7 @@ public:
     // ==========================================================================================================================
     // Statistics
     // ==========================================================================================================================
-    virtual QList<QPair<QList<QVariant>, long> > countItems(const QString &levelID, const QStringList &groupByAttributeIDs) = 0;
+    virtual QList<QPair<QList<QVariant>, long long> > countItems(const QString &levelID, const QStringList &groupByAttributeIDs) = 0;
 
 };
