+++ pncore/serialisers/sql/SQLAnnotationDatastore.cpp
@@ -318,7 +318,7 @@ bool SQLAnnotationDatastore::batchUpdate(const QString &levelID, const QString &
 // Statistics
 // ==========================================================================================================================
 
-QList<QPair<QList<QVariant>, long> > SQLAnnotationDatastore::countItems(const QString &levelID, const QStringList &groupByAttributeIDs)
+QList<QPair<QList<QVariant>, long long> > SQLAnnotationDatastore::countItems(const QString &levelID, const QStringList &groupByAttributeIDs)
 {
     return SQLSerialiserAnnotation::countItems(levelID, groupByAttributeIDs, d->structure, d->database);
 }
