+++ pncore/serialisers/sql/SQLAnnotationDatastore.h
@@ -102,7 +102,7 @@ public:
     // ==========================================================================================================================
     // Statistics
     // ==========================================================================================================================
-    QList<QPair<QList<QVariant>, long> > countItems(const QString &levelID, const QStringList &groupByAttributeIDs) override;
+    QList<QPair<QList<QVariant>, long long> > countItems(const QString &levelID, const QStringList &groupByAttributeIDs) override;
 
 private:
     SQLAnnotationDatastoreData *d;
