+++ pncore/serialisers/sql/SQLSerialiserAnnotation.h
@@ -82,8 +82,8 @@ public:
     // ==========================================================================================================================
     // Statistics
     // ==========================================================================================================================
-    static QList<QPair<QList<QVariant>, long> > countItems(const QString &levelID, const QStringList &groupByAttributeIDs,
-                                                           AnnotationStructure *structure, QSqlDatabase &db);
+    static QList<QPair<QList<QVariant>, long long> > countItems(
+            const QString &levelID, const QStringList &groupByAttributeIDs, AnnotationStructure *structure, QSqlDatabase &db);
 
 private:
     SQLSerialiserAnnotation();
